+++ src/OpenTelemetry.Instrumentation.EntityFrameworkCore/CHANGELOG.md
@@ -10,6 +10,9 @@
   repeated unterminated bracketed identifiers.
   ([#4339](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4339))
 
+* Fix `db.query.parameter.<key>` attributes to always emit the value as a string.
+  ([#TODO](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/TODO)
+
 ## 1.15.1-beta.1
 
 Released 2026-Apr-21