+++ src/OpenTelemetry.Instrumentation.SqlClient/CHANGELOG.md
@@ -17,6 +17,9 @@
   repeated unterminated bracketed identifiers.
   ([#4339](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4339))
 
+* Fix `db.query.parameter.<key>` attributes to always emit the value as a string.
+  ([#TODO](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/TODO)
+
 ## 1.15.2
 
 Released 2026-Apr-21